// src/modules/readings/domain/models/ReadingImages.ts
export interface ReadingImages {
  cadastralKey: string;
  readingId: number;
  previewsReading: number;
  currentReading: number;
  images: string[];
  readingMonth: string;
  readingYear: number;
  readingMonthName: string;
  novelty: string;
  consumption: number;
  observation: string;
  updatedStatus: boolean;
}